+++ src/Pim/Bundle/CatalogBundle/Entity/Repository/AttributeRepository.php
@@ -4,6 +4,7 @@ namespace Pim\Bundle\CatalogBundle\Entity\Repository;
 
 use Pim\Bundle\FlexibleEntityBundle\Entity\Repository\AttributeRepository as FlexibleAttributeRepository;
 use Pim\Bundle\EnrichBundle\Form\DataTransformer\ChoicesProviderInterface;
+use Doctrine\ORM\AbstractQuery;
 
 /**
  * Repository for attribute entity
@@ -254,6 +255,85 @@ class AttributeRepository extends FlexibleAttributeRepository implements
     public function getReferenceProperties()
     {
         return array('code');
+    }
+
+    /**
+     * Get attribute as array indexed by code
+     *
+     * @param boolean $withLabel translated label should be joined
+     * @param string  $locale    the locale code of the label
+     * @param array   $ids       the attribute ids
+     *
+     * @return array
+     */
+    public function getAttributesAsArray($withLabel = false, $locale = null, array $ids = [])
+    {
+        $qb = $this->_em->createQueryBuilder()
+            ->select('att')
+            ->from($this->_entityName, 'att', 'att.code');
+        if (!empty($ids)) {
+            $qb->andWhere('att.id IN (:ids)')->setParameter('ids', $ids);
+        }
+        $results = $qb->getQuery()->execute(array(), AbstractQuery::HYDRATE_ARRAY);
+
+        if ($withLabel) {
+            $labelExpr = 'COALESCE(trans.label, CONCAT(\'[\', att.code, \']\'))';
+            $qb = $this->_em->createQueryBuilder()
+                ->select('att.code', sprintf('%s as label', $labelExpr))
+                ->from($this->_entityName, 'att', 'att.code')
+                ->leftJoin('att.translations', 'trans', 'WITH', 'trans.locale = :locale')
+                ->setParameter('locale', $locale);
+            if (!empty($ids)) {
+                $qb->andWhere('att.id IN (:ids)')->setParameter('ids', $ids);
+            }
+            $labels = $qb->getQuery()->execute(array(), AbstractQuery::HYDRATE_ARRAY);
+            foreach ($labels as $code => $data) {
+                $results[$code]['label']= $data['label'];
+            }
+        }
+
+        return $results;
+    }
+
+    /**
+     * Get ids of attributes useable in grid
+     *
+     * @return array
+     */
+    public function getAttributeIdsUseableInGrid()
+    {
+        $qb = $this->_em->createQueryBuilder()
+            ->select('att.id')
+            ->from($this->_entityName, 'att', 'att.id');
+
+        $qb->andWhere(
+            "att.useableAsGridColumn = 1 ".
+            "OR att.useableAsGridFilter = 1 ".
+            "OR att.attributeType = 'pim_catalog_simpleselect'"
+        );
+        $result = $qb->getQuery()->execute([], AbstractQuery::HYDRATE_ARRAY);
+
+        return array_keys($result);
+    }
+
+    /**
+     * Get ids from codes
+     *
+     * @param mixed $codes the attribute codes
+     *
+     * @return array
+     */
+    public function getAttributeIds($codes)
+    {
+        $qb = $this->_em->createQueryBuilder()
+            ->select('att.id')
+            ->from($this->_entityName, 'att', 'att.id')
+            ->andWhere('att.code IN (:codes)');
+
+        $parameters = ['codes' => $codes];
+        $result = $qb->getQuery()->execute($parameters, AbstractQuery::HYDRATE_ARRAY);
+
+        return array_keys($result);
     }
 
     /**