+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/AttributeType/AttributeTypeTestCase.php
@@ -80,7 +80,7 @@ abstract class AttributeTypeTestCase extends \PHPUnit_Framework_TestCase
         if (!$this->target) {
             throw new \Exception(sprintf('You must override the setUp() method and provide a $target instance.'));
         }
-        $this->assertInstanceOf('Pim\Bundle\FlexibleEntityBundle\AttributeType\AbstractAttributeType', $this->target);
+        $this->assertInstanceOf('Pim\Bundle\CatalogBundle\AttributeType\AbstractAttributeType', $this->target);
     }
 
     /**