+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Validator/ConstraintGuesser/UniqueValueGuesserTest.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Tests\Unit\Validator\ConstraintGuesser;
 
-use Pim\Bundle\FlexibleEntityBundle\AttributeType\AbstractAttributeType;
+use Pim\Bundle\CatalogBundle\AttributeType\AbstractAttributeType;
 use Pim\Bundle\CatalogBundle\Validator\ConstraintGuesser\UniqueValueGuesser;
 
 /**