+++ src/Pim/Bundle/FlexibleEntityBundle/AttributeType/MetricType.php
@@ -5,6 +5,7 @@ namespace Pim\Bundle\FlexibleEntityBundle\AttributeType;
 use Pim\Bundle\FlexibleEntityBundle\Model\FlexibleValueInterface;
 use Akeneo\Bundle\MeasureBundle\Manager\MeasureManager;
 use Pim\Bundle\FlexibleEntityBundle\Form\Validator\ConstraintGuesserInterface;
+use Pim\Bundle\CatalogBundle\AttributeType\AbstractAttributeType as NewAbstractAttributeType;
 
 /**
  * Metric attribute type
@@ -12,8 +13,10 @@ use Pim\Bundle\FlexibleEntityBundle\Form\Validator\ConstraintGuesserInterface;
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @deprecated Deprecated since version 1.1, to be removed in 1.2. Use CatalogBundle/AttributeType
  */
-class MetricType extends AbstractAttributeType
+class MetricType extends NewAbstractAttributeType
 {
     /**
      * @var MeasureManager $manager