+++ src/Pim/Bundle/FlexibleEntityBundle/Tests/Unit/Form/Validator/AttributeConstraintGuesserTest.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\FlexibleEntityBundle\Tests\Unit\Form\Validator;
 
-use Pim\Bundle\FlexibleEntityBundle\AttributeType\AbstractAttributeType;
+use Pim\Bundle\CatalogBundle\AttributeType\AbstractAttributeType;
 use Symfony\Component\Validator\Constraints;
 use Pim\Bundle\FlexibleEntityBundle\Form\Validator\AttributeConstraintGuesser;