+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Changes in the implementation of storing datagrid state - adding 'pim/datagrid/state-listener' to the datagrid configuration is no longer required, instead, the grid should be rendered with dataGrid.renderStatefulGrid()
 - Move `FilterBundle/Filter/Flexible/*` -> `FilterBundle/Filter/ProductValue/*`
 - Remove unused FilterBundle/Filter/ProductValue/EntityFilter
+- Replace FlexibleManager by ProductManager in ContextConfigurator constructor arguments
 
 # 1.0.1
 ## Bug Fixes